avfilter: split individual filters to separate files
Splits the initialization of pad, cropscale, deinterlace, rotate, and
colorspace into their own files.  Also splits internal avfilter API into
it's own file.
